// src/phamarcy/DTOs/Utils/types.ts
export type CreatePhamarcyParams={
    DrugName: string;
    DrugType: string;
    Quantity: number;
    Amount: number;
    MedicalScheme: string;
    expiryDate: string
}
export class  UpdatedPhamarcyParams {
    DrugName: string;
    DrugType: string;
    Quantity: number;
    Amount: number;
    MedicalScheme: string;
    expiryDate: string
 
  }